checkAuth & checkLogin moved to authService
