+++ frontend/src/services/auth.js
@@ -1,5 +1,5 @@
 import axios from 'axios'
-
+import { redirect } from 'react-router-dom'
 
 const baseUrl = process.env.REACT_APP_BASEURL
 
@@ -17,4 +17,28 @@ const getUser = async token => {
   }
 }
 
-export default { login, getUser }
+const checkAuth = async () => {
+  const token = localStorage.getItem('token')
+  if (!token) {
+    return redirect('/login', { replace: true })
+  }
+  const user = await getUser(token)
+  if (user.status && user.status !== 200) {
+    localStorage.removeItem('token')
+    return redirect('/login', { replace: true })
+  }
+  return null
+}
+
+const checkLogin = async () => {
+  const token = localStorage.getItem('token')
+  if (token) {
+    const user = await getUser(token)
+    if (user.status && user.status === 200) {
+      return redirect('/', { replace: true })
+    }
+  }
+  return null
+}
+
+export default { login, getUser, checkAuth, checkLogin }
